turn hyperparam searching into function
